Changes: Ut_MOrientationTracker - get compared values when QCOMPARE fails
RevBy: TrustMe
Details: Cast then to int (a type known by QCOMPARE) otherwise a QCOMPARE
failure won't print the actual and expected values compared.

+++ tests/ut_morientationtracker/ut_morientationtracker.cpp
@@ -325,8 +325,8 @@ void Ut_MOrientationTracker::testWindowRemoteOrientationLock()
 
     QCoreApplication::processEvents();
 
-    QCOMPARE(window1->orientationAngle(), expectedAngle);
-    QCOMPARE(window2->orientationAngle(), expectedAngle);
+    QCOMPARE(static_cast<int>(window1->orientationAngle()), static_cast<int>(expectedAngle));
+    QCOMPARE(static_cast<int>(window2->orientationAngle()), static_cast<int>(expectedAngle));
 }
 
 ///////////////////////////////////////////////////////
